fix) jsx-> js svgcomponent -> Image
