+++ styles/globals.css
@@ -403,3 +403,6 @@ template {
     margin: 0 auto;
   }
 }
+html {
+  font-size: 16px;
+}
